Rename /test as /check
