+++ app/plugins/router.plugin.js
@@ -13,10 +13,10 @@
 
 const AssetRoutes = require('../routes/assets.routes.js')
 const BillRunRoutes = require('../routes/bill-runs.routes')
+const CheckRoutes = require('../routes/check.routes.js')
 const FilterRoutesService = require('../services/plugins/filter-routes.service.js')
 const HealthRoutes = require('../routes/health.routes.js')
 const RootRoutes = require('../routes/root.routes.js')
-const TestRoutes = require('../routes/test.routes.js')
 
 const AirbrakeConfig = require('../../config/airbrake.config.js')
 
@@ -24,7 +24,7 @@ const routes = [
   ...RootRoutes,
   ...AssetRoutes,
   ...HealthRoutes,
-  ...TestRoutes,
+  ...CheckRoutes,
   ...BillRunRoutes
 ]
