+++ app/routes/check.routes.js
@@ -1,0 +1,16 @@
+'use strict'
+
+const SupplementaryController = require('../controllers/check/supplementary.controller.js')
+
+const routes = [
+  {
+    method: 'GET',
+    path: '/check/supplementary',
+    handler: SupplementaryController.index,
+    options: {
+      description: 'Endpoint to check the SROC supplementary billing logic'
+    }
+  }
+]
+
+module.exports = routes
